Details Plan Conclused

+++ app/Http/Controllers/Admin/PlanController.php
@@ -39,7 +39,7 @@ class PlanController extends Controller
 
     public function show($url)
     {
-        
+
         $plan = $this->repository->where('url', $url)->first();
 
         if(!$plan) {
@@ -53,10 +53,18 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                            ->with('details')
+                            ->where('url', $url)->first();
 
         if(!$plan) {
             return redirect()->back();
+        }
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                            ->back()
+                            ->with('error', 'Existem detalhes vinculados a esse plano, favor deletá-los antes!');
         }
 
         $plan->delete();
